+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - Added optional parallax effect for hero image
+- Added style option for top, main and bottom sections
 - Added optional off-canvas logo
 
 ## 0.10.0